+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## UNRELEASED
+## 0.3.0 (UNRELEASED)
 
 - [Improvement] Don't return an environment that has no versions of the cookbook in it.
 - [Feature] Added new `bump` command which will bump the version of the current cookbook.
+- [Feature] Added new `upload` command which will upload the current cookbook.
 
 ## 0.2.0